Old messages are no longer repeated when, for example, creating a new project.
This repeating occurred because we tweak the name of the focus and REAPER reuses the same window for every project. Therefore, even when dealing with a different project, the old name tweak still applied.
The name tweak is now cleared when the focus changes.
This doesn't cover cases where there is no focus change for a project switch; e.g. switching project tabs. However, we don't get any reporting at all in that case, so taht needs ot be dealt with separately.
Fixes #17.