Make report cursor position and shortcut help work in the MIDI Event List.
The previous trick of tweaking the name of the focus doesn't work in the Event List because of child ids and because NVDA (and probably other screen readers) ignore MSAA for ListViews.
In order to do this, outputMessage now creates a fake focused accessible and uses that to output messages.
Fixes #56 and fixes #57.