You can now select envelopes for the focused track or take.
This is done with the new actions: OSARA: Select next/previous track/take envelope (depending on focus).
If the envelope is bypassed or armed, this will be reported.
Support for Track: Select next/previous envelope has been removed, as these actions include envelopes for all tracks, not just the current one. Also, this code was necessarily pretty inefficient and ugly.