Add files to build a Mac installer package (.pkg).
Unfortunately, this isn't going to work for users. OSARA *must* be installed to the user's home directory because of the key map. However, there's no way to force the Mac installer to install to the home directory by default, even when the anywhere and localSystem domains are disabled in the package distribution file. Worse, the radio buttons to change the install location aren't labelled properly for VoiceOver.
Still, it's worth keeping this code for reference and maybe future use.