Fix accuracy/consistency of the Peak Watcher.
Previously, we retrieved the peak info for the track every 30 ms using Track_GetPeakInfo, but this meant we could miss peaks, causing inaccuracy.
The inconsistency occurred because the time at which we fetched peaks differed depending on where we were at in our 30 ms interval when playback started.
To fix this, use Track_GetPeakHoldDB, which holds the peak until explicitly cleared.
I missed this API function when I originally wrote this code.
Fixes #119.