Fix stretch marker functionality when a take's play rate is adjusted.
Stretch marker positions are relative to the audio within the take.
That means we have to subtract the item start position (position 0 is the start of the item), which we were already doing.
However, we must also multiply by the play rate (1 second in real time is 2 seconds of take audio).
Fixes #161.