Add actions to show REAPER context menus depending on focus.
On Windows, we show these menus by intercepting the applications key, since that's the native way to access context menus on Windows.
On the Mac, there's no native way to do this and thus nothing we can override.
Therefore, provide actions that can be used on the Mac.
These aren't bound to keys yet.