Parameters dialogs: Don't allow the slider to be adjusted beyond the minimum or maximum parameter value. Don't try to snap when the value is the same.
Previously, the range was being checked when snapping to the next value, but not for the initial increment/decrement.
Also, when pressing home/end, we were trying to snap even if the value was the same; e.g. if the user pressed end and then end again.
Now, if the initial new value is the same or outside of min/max, we just return early.
Fixes #209.