Parameters dialogs: Report percentages for the slider value if formatted (friendly) values are not supported.
Previously, we relied on the slider control's own percentage reporting if there was no formatted value.
Now that we don't actually increment the slider control's value (it's effectively just a placeholder), this doesn't work.
Instead, when there's no formatted value, fall back to calculating the percentage ourselves.
This has the added benefit of allowing us to report values in tenths of a percent, which is more accurate since we allow 1000 steps for FX values.
Fixes #208.