Fix actions for opening I/O window for master track in REAPER >= 6.01.
The track accessibility implementation has changed in REAPER 6, so the technique used for REAPER 5 no longer works.
However, that technique also broke if the user used a different theme which moved the I/O button.
REAPER 6.01 adds a builtin action for this, so use that instead.
Eventually, once we no longer need to support REAPER 5, we can remove "OSARA: View I/O for master track" altogether.

This also means that for REAPER >= 6.01, actions to open master I/O now work on Mac as well.

Fixes #232. Fixes #50.