Report whether the selected effect in FX Chain dialogs is active or bypassed. Fixes #267.
We can't annotate the names of SysListView32 items, since screen readers have special support for those and override MSAA.
So, the user has to tab to the notes field (one tab stop from the FX list) to have the active/bypassed state reported.
The change will also be reported if the user presses control+b while focused in that field.
This isn't ideal and is a big ugly hack, but it's far better than nothing and is the best I could come up with for now.