Fix potential crashes/bogus data when inserting a MIDI note.
The insert node command behaves very differently in the Event List.
It brings up a dialog which allows the user to change the event to a different type; e.g. CC.
In addition, the pitch can be changed and it doesn't default to the pitch cursor in the other MIDI views.
Therefore, don't try to use cmdMidiInsertNote at all in the Event List.

That said, I took this opportunity to robustify cmdMidiInsertNote anyway:

1. Count notes, not events, since the command might add some other kind of event.
2. Don't assume a note matching the pitch cursor was inserted without checking, as this can cause us to dereference an invalid pointer.

Fixes #428.