Allow access to context menus using shift+f10.
For track view context menus, we still need to support the applications key variants using the keyboard hook, as they don't work properly even when bound to actions.
The shift+f10 variants are supported using action bindings.
Because alt+shift+f10 is already bound, the third context menu is accessed using control+alt+shift+f10.

It also turns out control+applications wasn't working properly and was bringing up the wrong context menu.
It seems we can't override control+applications on key up for some reason as we were previously trying to do, but we can get around this by bringing up the menu on key down.

The keyboard hook was refactored a bit to support shift+f10 for the OSARA send menu in the I/O window and to simplify the code.

Fixes #432.