Fix crash when there is an error in a translated format string.
If translators accidentally modify brace placeholders, this can cause the format string to be invalid.
In that case, fmt::format will throw an exception, causing a crash.
Ideally, we want to catch this before shipping broken translations to users, but we shouldn't crash regardless.

To fix this, create our own format() function which catches the exception and returns a message informing of the problem.
As part of this, all code related to translation was moved from osara.h to a new translation.h.
Also, the format argument to formatTime() in reaper_osara.cpp had to be renamed to avoid shadowing OSARA's format() function, since it isn't namespaced.

Fixes #491.