Fix extremely short pitch cursor previews in the MIDI editor.
If a length was specified, we were multiplying that by 1000 to get ms.
However, where a length wasn't specified, the default time was in seconds, but we weren't multiplying that, resulting in near 0 length previews.

As part of this, don't restrict the preview length to a minimum of the default, as some notes might be deliberately shorter.
Instead, only use the default length if no length is specified.

Fixes #531.