Catch possible exception when trying to determine the note for MIDI event list preview. (issue #545, PR #554)
It seems that REAPER language packs can translate note names.
If a translated note name matches the start of an English note name (e.g. "Fa" will match "F"), stoi (used to get the octave) will then barf because "a" isn't a number.
For now, just catch this exception and gracefully ignore it to prevent a crash.
In future, we're going to have to figure out how to deal with translated note names.