Rewrite Peak Watcher to support multiple level types and watching things other than tracks.
Instead of choosing two tracks within the dialog, the currently focused track is used.
You are then asked to choose which of two watchers you want to configure.
If either of these watchers has been configured already, the menu provides information about the existing configuration as a reminder.

In this commit, only tracks are supported, but the design allows for other things such as track FX in future commits.