Provide map to simplify messages for toggle actions and report the state of toggle actions not explicitly supported by OSARA. (#712)
The TOGGLE_COMMAND_MESSAGES map makes it much simpler to define on and off messages for toggle actions with a single line of code.
A null message can be specified if nothing should be reported.
For toggle actions which aren't included in any of OSARA's maps, OSARA will report "enabled" or "disabled", followed by the action name.