Avoid unintended FX parameter change
In FX parameters dialog, edit control always set parameter of focus kill. Since precision of displayed number is fixed to 4, that can change actual parameter value even in case the text was not modified.

For example in ReqEQ if band frequency was 1000 Hz it is changed to 999.7 Hz.

This change prevent parameter editing in case the text was not modified.