Avoid unintended parameter change in the Parameters dialog when tabbing away from the value text box. (#751)
The parameter value is always set when focus moves away from the text box. Since precision of displayed number is fixed to 4, that could previously change the actual parameter value even when the text was not modified.
For example, in ReaEQ, if band frequency was 1000 Hz, it was changed to 999.7 Hz.
This change prevents the parameter from being set if the text was not modified.