Refactor configuration handling to make it easier to maintain and extend.
Settings are now defined in settings.h.
Each setting is defined using a simple macro so that all info for the setting is in one place.
The build system uses this to generate the dialog GUI resource automatically.
It is also used to load and save the settings and to set and retrieve the right states in the GUI.
Code that wants to access the settings should include config.h.
Each setting is a bool variable in the settings namespace; e.g. settings::reportScrub.