Fix take volume in Parameters dialog when take polarity is flipped.
In this case, the raw value is negative.
We need to flip it accordingly when both retrieving and setting it.
It was easiest to do this in ReaperObjVolParam, even though this only affects takes.
We check whether the current raw value is negative during construction and set a flag so we know whether the sign needs to be flipped thereafter.
This avoids checks each time we set the value; i.e. we only check this the first time the parameter is constructed.
Fixes #891.