Automatically check for updates.
OSARA will check for an update when REAPER starts if it hasn't checked within the last 24 hours.
You can also manually check for updates using the action: OSARA: Check for update.
If an update is available, the dialog will show the commit log since you last updated.
OSARA can't download and install the update, so it provides a button to go directly to the download page.
Fixes #619.