Don't try to use UIA before we've even tried to initialize it.
REAPER can sometimes cause OSARA to report a message very early in startup before we've tried to initialize UIA.
When this happened, we previously fell back to MSAA for the entire session.
This broke OSARA's ability to interrupt messages and report when a ListView is focused, among other things.
To fix this, if we haven't tried to initialize UIA yet, just ignore the message.
A message this early isn't useful anyway.

Fixes #1173.