Don't provide feedback for SWS actions if SWS is not installed.
If an extension action is bound in the key map, it seems REAPER registers a numeric id for it even if the extension isn't running.
Previously, we intercepted that id such that OSARA's code for that action ran, even though the action didn't exist.
Now, we check if there is a name registered for the action.
If there isn't, we know the extension isn't installed and we thus ignore the action.
Fixes #1202.