ci: Switch from AppVeyor to GitHub Actions for automated builds. Use GitHub Pages for the website.
Almost everything is on GitHub now, including snapshots, pull request builds, the website and all the automation thereof.
Snapshots are uploaded as GitHub release artifacts so they can be downloaded directly.
Pull request builds are uploaded as normal GitHub artifacts.
The website is built and deployed to GitHub pages as part of the build automation, including the snapshots page and the update check data.
See the comments in the GitHub workflow code for more details.
As well as standardising on a single platform, this removes the dependency on my personal server and AppVeyor account, which makes things more sustainable and easier for others to collaborate on.