Fix broken menu items in the Extensions -> OSARA menu on Mac.
In #1180, OSARA started registering all commands shortly after startup instead of immediately on startup.
This was done to enable interception of SWS commands.
Unfortunately, this broke items in the OSARA menu on Mac.
Presumably, the menu hook gets called slightly earlier on Mac, before our commands get registered, and so our menu items can't be associated with the correct command.
To fix this, we now register OSARA's commands immediately on startup, but we continue to register other commands that we want to intercept slightly later.
Developers should note that OSARA's own commands are now listed in the OSARA_COMMANDS array, not the COMMANDS array.