CI: Improve Windows build efficiency on GitHub Actions by caching the SCons MSVC configuration (#1235)
Executing the SCons script reading phase on Windows GitHub Actions can take more than a minute, because calculating MSVC build variables is very slow. This change ensures that we cache the MSVC setup bound to the OS and image version. This means that the cache will be refreshed properly when the Actions runner image is updated.