New OSARA action names and keystrokes in MIDI Editor. (#1258) (Fixes #1251).
The OSARA actions for moving to previous and next notes in chord have been renamed, now called move to lower or higher note in chord for clarity. 
MIDI Editor Key Map Improvements: 
OSARA: Move to higher note in chord: UpArrow (with Shift to add to selection). 
OSARA: Move to lower note in chord: DownArrow (with Shift to add to selection).